requirements.txt
streamlit
google-genai
google-generativeai
Pillow
gTTS